remove static method in os (why was it there in the first place?) regular #if breaks on msvc / use ifdef?
