+++ Src/Prodos_Extract.c
@@ -10,7 +10,7 @@
 #include <stdlib.h>
 #include <string.h>
 
-#if IS_WINDOWS
+#ifdef IS_WINDOWS
 #include <malloc.h>
 #endif
 
@@ -313,7 +313,7 @@ static int CreateOutputFile(struct prodos_file *current_file, char *output_direc
   /** Ajustement des Dates **/
   os_SetFileCreationModificationDate(file_data_path,current_file->entry);
 
-  #if IS_WINDOWS
+  #ifdef IS_WINDOWS
   /** Change la visibilité du fichier **/
   os_SetFileAttribute(
     file_data_path, ((current_file->entry->access & 0x04) == 0x04) ? SET_FILE_HIDDEN : SET_FILE_VISIBLE
@@ -340,7 +340,7 @@ static int CreateOutputFile(struct prodos_file *current_file, char *output_direc
       /** Ajustement des Dates **/
       os_SetFileCreationModificationDate(file_resource_path,current_file->entry);
 
-      #if IS_WINDOWS
+      #ifdef IS_WINDOWS
       /** Change la visibilité du fichier **/
       os_SetFileAttribute(
         file_resource_path,
@@ -387,14 +387,14 @@ static void SetFileInformation(char *file_information_path, struct prodos_file *
       /* Créer le fichier FileInformation */
       CreateBinaryFile(file_information_path,(unsigned char *)local_buffer,(int)strlen(local_buffer));
 
-      #if IS_WINDOWS
+      #ifdef IS_WINDOWS
       /* Rendre le fichier invisible */
       os_SetFileAttribute(file_information_path, SET_FILE_HIDDEN);
       return;
       #endif
     }
 
-  #if IS_WINDOWS
+  #ifdef IS_WINDOWS
   /* Rendre le fichier visible */
   os_SetFileAttribute(file_information_path, SET_FILE_VISIBLE);
   #endif
@@ -434,7 +434,7 @@ static void SetFileInformation(char *file_information_path, struct prodos_file *
   mem_free_list(nb_line,line_tab);
 
   /* Rendre le fichier invisible */
-  #if IS_WINDOWS
+  #ifdef IS_WINDOWS
   os_SetFileAttribute(file_information_path, SET_FILE_HIDDEN);
   #endif
 }
